Create new config file and code to process it
The config file is in a new directory called input. The items are all left blank currently. The read_input.cpp file is incomplete and there are still some things to work out with the yaml-cpp library, especially it's connection in CMake. The commented areas will be filled in when we have both the Room and Robot classes ready and they can be instantiated properly.